requirements.txt
google-adk==1.20.0
google-auth==2.43.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.1
google-api-python-client==2.187.0
google-genai==1.53.0
python-dotenv==1.2.1
requests==2.31.0
pypdf==6.4.1
python-docx
openpyxl
